requirements.txt
--extra-index-url https://testpypi.python.org/pypi
grakn-kglib
absl-py==0.5.0
astor==0.7.1
decorator==4.3.0
gast==0.2.0
http://repo.grakn.ai/repository/test-pypi-group/packages/grakn-client/1.5.0-02ccd4a4945924a5f81e7aeb7b0b3fa1fd7e7fe2/grakn-client-1.5.0-02ccd4a4945924a5f81e7aeb7b0b3fa1fd7e7fe2.tar.gz
grpcio==1.15.0
h5py==2.8.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
networkx==2.2
numpy==1.15.2
protobuf==3.6.1
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
tensorboard==1.11.0
tensorflow==1.11.0
tensorflow-hub==0.1.1
termcolor==1.1.0
Werkzeug==0.14.1
